typescript: Convert Conversation model
