+++ server/graphql/common/comment.ts
@@ -4,13 +4,14 @@ import ActivityTypes from '../../constants/activities';
 import { mustBeLoggedInTo } from '../../lib/auth';
 import models from '../../models';
 import Comment from '../../models/Comment';
+import Conversation from '../../models/Conversation';
 import { NotFound, Unauthorized, ValidationFailed } from '../errors';
 
 import { canComment } from './expenses';
 import { checkRemoteUserCanUseComment } from './scope-check';
 import { canSeeUpdate } from './update';
 
-type CommentableEntity = typeof models.Update | typeof models.Expense | typeof models.Conversation;
+type CommentableEntity = typeof models.Update | typeof models.Expense | Conversation;
 
 const loadCommentedEntity = async (commentValues): Promise<[CommentableEntity, ActivityTypes]> => {
   const include = { association: 'collective', required: true };
